chore: move mocks lib into folder and improve reports mock lib
