+++ packages/bindings/src/mocks/mock_app.rs
@@ -963,7 +963,7 @@ mod tests {
             msg::RelationshipsMsg, querier::RelationshipsQuerier,
         },
         reports::{
-            mocks::get_mocked_report, models_query::QueryReportResponse, msg::ReportsMsg,
+            mocks::MockReportsQueries, models_query::QueryReportResponse, msg::ReportsMsg,
             querier::ReportsQuerier,
         },
         subspaces::{
@@ -1121,7 +1121,7 @@ mod tests {
         let querier = ReportsQuerier::new(app_querier.deref());
         let response = querier.query_report(1, 1).unwrap();
         let expected = QueryReportResponse {
-            report: get_mocked_report(&1u64.into()),
+            report: MockReportsQueries::get_mocked_report(&1u64.into()),
         };
         assert_eq!(expected, response)
     }
