+++ packages/bindings/src/reactions/querier.rs
@@ -216,7 +216,7 @@ impl<'a> ReactionsQuerier<'a> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::mocks::mock_dependencies_with_custom_querier;
+    use crate::mocks::mock_queriers::mock_dependencies_with_custom_querier;
     use crate::reactions::mocks::MockReactionsQueries;
     use std::ops::Deref;
